Begin refactoring: code and structure
