--- .travis.yml
@@ -1,7 +1,0 @@
-language: python
-python:
-  - "2.7"
-install:
-  - pip install -r requirements.txt
-
-script: python src/gui/maprender.py
