+++ README.md
@@ -1,4 +1,3 @@
 # Stargate RL
 ###### A Stargate inspired RogueLike
-
 ---
