--- bin/.gitignore
@@ -1,1 +1,0 @@
-!colorlist.pkl.zip
