+++ requirements.txt
@@ -1,1 +1,1 @@
-pyglet==1.2.4
+pyglet
