--- src/manager/game_ui.py~
@@ -1,9 +1,0 @@
-from utility import config
-
-class FrameUI:
-    
-    def __init__ (self):
-        pass
-
-    def update_size(self):
-        pass
